+++ source/renderer/app/Routes.js
@@ -19,10 +19,11 @@ import DisplaySettingsPage from './containers/settings/categories/DisplaySetting
 import PaperWalletCreateCertificatePage from './containers/wallet/PaperWalletCreateCertificatePage';
 import Staking from './containers/staking/Staking';
 import StakingDelegationCenterPage from './containers/staking/StakingDelegationCenterPage';
-import StakingEpochsPagePage from './containers/staking/StakingEpochsPage';
+import StakingEpochsPage from './containers/staking/StakingEpochsPage';
 import StakingInfoPage from './containers/staking/StakingInfoPage';
 import StakingRewardsPage from './containers/staking/StakingRewardsPage';
 import StakingStakePoolsPage from './containers/staking/StakingStakePoolsPage';
+import StakingDelegationCountdownPage from './containers/staking/StakingDelegationCountdownPage';
 import Wallet from './containers/wallet/Wallet';
 import WalletSummaryPage from './containers/wallet/WalletSummaryPage';
 import WalletSendPage from './containers/wallet/WalletSendPage';
@@ -75,7 +76,11 @@ export const Routes = (
       component={PaperWalletCreateCertificatePage}
     />
     <Route path={ROUTES.STAKING.ROOT} component={Staking}>
-      <IndexRedirect to={ROUTES.STAKING.DELEGATION_CENTER} />
+      <IndexRedirect to={ROUTES.STAKING.INFO} />
+      <Route
+        path={ROUTES.STAKING.DELEGATION_COUNTDOWN}
+        component={StakingDelegationCountdownPage}
+      />
       <Route
         path={ROUTES.STAKING.DELEGATION_CENTER}
         component={StakingDelegationCenterPage}
@@ -85,7 +90,7 @@ export const Routes = (
         component={StakingStakePoolsPage}
       />
       <Route path={ROUTES.STAKING.REWARDS} component={StakingRewardsPage} />
-      <Route path={ROUTES.STAKING.EPOCHS} component={StakingEpochsPagePage} />
+      <Route path={ROUTES.STAKING.EPOCHS} component={StakingEpochsPage} />
       <Route path={ROUTES.STAKING.INFO} component={StakingInfoPage} />
     </Route>
   </Route>