+++ source/renderer/app/actions/staking-actions.js
@@ -1,0 +1,8 @@
+// @flow
+import Action from './lib/Action';
+
+// ======= STAKING ACTIONS =======
+
+export default class StakingActions {
+  goToStakingPage: Action<any> = new Action();
+}