+++ source/renderer/app/config/sidebarConfig.js
@@ -3,7 +3,8 @@ import { ROUTES } from '../routes-config';
 import walletsIcon from '../assets/images/sidebar/wallet-ic.inline.svg';
 import settingsIcon from '../assets/images/sidebar/settings-ic.inline.svg';
 import paperWalletCertificateIcon from '../assets/images/sidebar/paper-certificate-ic.inline.svg';
-import stakingIcon from '../assets/images/sidebar/delegation-ic.inline.svg';
+import delegationIcon from '../assets/images/sidebar/delegation-ic.inline.svg';
+import delegationProgressIcon from '../assets/images/sidebar/delegation-progress-ic.inline.svg';
 
 export const CATEGORIES_BY_NAME = {
   WALLETS: {
@@ -16,10 +17,15 @@ export const CATEGORIES_BY_NAME = {
     route: ROUTES.PAPER_WALLET_CREATE_CERTIFICATE,
     icon: paperWalletCertificateIcon,
   },
+  STAKING_DELEGATION_COUNTDOWN: {
+    name: 'STAKING_DELEGATION_COUNTDOWN',
+    route: ROUTES.STAKING.DELEGATION_COUNTDOWN,
+    icon: delegationIcon,
+  },
   STAKING: {
     name: 'STAKING',
     route: ROUTES.STAKING.ROOT,
-    icon: stakingIcon,
+    icon: delegationProgressIcon,
   },
   SETTINGS: {
     name: 'SETTINGS',
@@ -34,4 +40,31 @@ export const CATEGORIES = [
   CATEGORIES_BY_NAME.SETTINGS,
 ];
 
-export const sidebarConfig = { CATEGORIES_BY_NAME, CATEGORIES };
+export const CATEGORIES_WITH_DELEGATION_COUNTDOWN = [
+  CATEGORIES_BY_NAME.WALLETS,
+  CATEGORIES_BY_NAME.PAPER_WALLET_CREATE_CERTIFICATE,
+  CATEGORIES_BY_NAME.STAKING_DELEGATION_COUNTDOWN,
+  CATEGORIES_BY_NAME.SETTINGS,
+];
+
+export const CATEGORIES_WITHOUT_DELEGATION_COUNTDOWN = [
+  CATEGORIES_BY_NAME.WALLETS,
+  CATEGORIES_BY_NAME.PAPER_WALLET_CREATE_CERTIFICATE,
+  CATEGORIES_BY_NAME.STAKING,
+  CATEGORIES_BY_NAME.SETTINGS,
+];
+
+export const CATEGORIES_WITH_STAKING = [
+  CATEGORIES_BY_NAME.WALLETS,
+  CATEGORIES_BY_NAME.STAKING_DELEGATION_COUNTDOWN,
+  CATEGORIES_BY_NAME.STAKING,
+  CATEGORIES_BY_NAME.SETTINGS,
+];
+
+export const sidebarConfig = {
+  CATEGORIES_BY_NAME,
+  CATEGORIES,
+  CATEGORIES_WITH_DELEGATION_COUNTDOWN,
+  CATEGORIES_WITHOUT_DELEGATION_COUNTDOWN,
+  CATEGORIES_WITH_STAKING,
+};