+++ source/renderer/app/stores/AppStore.js
@@ -105,10 +105,6 @@ export default class AppStore extends Store {
     );
   }
 
-  @computed get isDelegationPage(): boolean {
-    return this.currentRoute === ROUTES.STAKING;
-  }
-
   // ===================== PRIVATE ======================= //
 
   _getGpuStatus = async () => {